add mimetype() to Path
