+++ src/util/mod.rs
@@ -1,11 +1,18 @@
 mod dir_entry;
 pub use self::dir_entry::DirEntryExt;
 mod path;
-pub use self::path::get_ext;
+pub use self::path::get_mimetype;
 pub use self::path::PathExt;
 
 use pulldown_cmark::CowStr;
-use std::path::Path;
+use std::{borrow::Cow, path::Path};
+
+// return the extension of an url as a lowercase string
+// or empty string, if there is no extension
+pub fn get_ext<T: AsRef<str>>(url: T) -> Cow<'static, str> {
+    let path = Path::new(url.as_ref());
+    path.get_ext()
+}
 
 pub fn is_audio_file(url: &CowStr) -> bool {
     let audio_format = ["mp3", "mp4", "m4a", "wav", "ogg"];
@@ -17,39 +24,6 @@ pub fn is_audio_file(url: &CowStr) -> bool {
         }
     }
     false
-}
-
-// return mimetype given an extension
-pub fn get_mimetype(ext: &str) -> String {
-    info!("get_mimetype for: {}", ext);
-    match ext {
-        "mp3" => "audio/mpeg",
-        "mp4" => "video/mp4",
-        "m4a" => "audio/mp4",
-        "wav" => "audio/wav",
-        "ogg" => "audio/ogg",
-        "jpg" => "image/jpeg",
-        "jpeg" => "image/jpeg",
-        "png" => "image/png",
-        "gif" => "image/gif",
-        "svg" => "image/svg+xml",
-        "webp" => "image/webp",
-        "pdf" => "application/pdf",
-        "zip" => "application/zip",
-        "gz" => "application/gzip",
-        "tar" => "application/x-tar",
-        "txt" => "text/plain",
-        "md" => "text/markdown",
-        "html" => "text/html",
-        "css" => "text/css",
-        "js" => "text/javascript",
-        "json" => "application/json",
-        "xml" => "application/xml",
-        "yaml" => "text/yaml",
-        "yml" => "text/yaml",
-        _ => "application/octet-stream",
-    }
-    .to_string()
 }
 
 #[cfg(test)]
@@ -66,15 +40,5 @@ mod tests {
     fn test_get_ext_empty() {
         let result = get_ext("");
         assert_eq!(result, "".to_string());
-    }
-    #[test]
-    fn test_get_mimetype_png() {
-        let result = get_mimetype(&CowStr::Borrowed("png"));
-        assert_eq!(result, "image/png".to_string());
-    }
-    #[test]
-    fn test_get_mimetype_empty() {
-        let result = get_mimetype(&CowStr::Borrowed(""));
-        assert_eq!(result, "application/octet-stream".to_string());
     }
 }
