+++ src/util/path.rs
@@ -1,11 +1,36 @@
 //-- Path utlity functions
 use std::{borrow::Cow, path::Path};
 
-// return the extension of an url as a lowercase string
-// or empty string, if there is no extension
-pub fn get_ext<T: AsRef<str>>(url: T) -> Cow<'static, str> {
-    let path = Path::new(url.as_ref());
-    path.get_ext()
+// return mimetype given an extension
+pub fn get_mimetype(ext: &str) -> Cow<'static, str> {
+    info!("get_mimetype for: {}", ext);
+    Cow::from(match ext {
+        "mp3" => "audio/mpeg",
+        "mp4" => "video/mp4",
+        "m4a" => "audio/mp4",
+        "wav" => "audio/wav",
+        "ogg" => "audio/ogg",
+        "jpg" => "image/jpeg",
+        "jpeg" => "image/jpeg",
+        "png" => "image/png",
+        "gif" => "image/gif",
+        "svg" => "image/svg+xml",
+        "webp" => "image/webp",
+        "pdf" => "application/pdf",
+        "zip" => "application/zip",
+        "gz" => "application/gzip",
+        "tar" => "application/x-tar",
+        "txt" => "text/plain",
+        "md" => "text/markdown",
+        "html" => "text/html",
+        "css" => "text/css",
+        "js" => "text/javascript",
+        "json" => "application/json",
+        "xml" => "application/xml",
+        "yaml" => "text/yaml",
+        "yml" => "text/yaml",
+        _ => "application/octet-stream",
+    })
 }
 
 pub trait PathExt {
@@ -13,6 +38,7 @@ pub trait PathExt {
     // and create any that don't exist
     fn create_all_parent_dir(&self) -> std::io::Result<()>;
     fn get_ext(&self) -> Cow<'static, str>;
+    fn mimetype(&self) -> Cow<'static, str>;
 }
 
 impl PathExt for Path {
@@ -30,5 +56,35 @@ impl PathExt for Path {
         } else {
             Cow::Borrowed("")
         }
+    }
+    fn mimetype(&self) -> Cow<'static, str> {
+        get_mimetype(&self.get_ext())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    // importing names from outer (for mod tests) scope.
+    use super::*;
+
+    #[test]
+    fn test_get_ext_png() {
+        let result = Path::new("foo.png").get_ext();
+        assert_eq!(result, "png".to_string());
+    }
+    #[test]
+    fn test_get_ext_empty() {
+        let result = Path::new("").get_ext();
+        assert_eq!(result, "".to_string());
+    }
+    #[test]
+    fn test_get_mimetype_png() {
+        let result = get_mimetype("png");
+        assert_eq!(result, "image/png".to_string());
+    }
+    #[test]
+    fn test_get_mimetype_empty() {
+        let result = get_mimetype("");
+        assert_eq!(result, "application/octet-stream".to_string());
     }
 }
